+++ hqporner_api/exceptions.py
@@ -1,0 +1,13 @@
+class InvalidCategory(Exception):
+    def __init__(self):
+        self.msg = "Invalid Category!"
+
+
+class NoVideosFound(Exception):
+    def __init__(self):
+        self.msg = "No videos were found!"
+
+
+class InvalidActress(Exception):
+    def __init__(self):
+        self.msg = "Invalid Actress!"