+++ README/CHANGELOG.md
@@ -8,3 +8,14 @@
 - Better documentation
 - Better readme
 - Removed exceptions
+
+# 1.0
+
+- Split video methods into a new `Video` object
+- Refactored code
+- You can now search for videos
+- You can now get videos by a category
+- Fixed the get videos by actress page issue
+- Added exceptions
+- Updated Documentation
+- Added a new progressbar